# public/openapi.yaml
openapi: 3.1.0
info:
  title: tevaxia.lu API
  version: 1.1.1
  description: >-
    API de calculs indicatifs. Les hypothèses internes ne constituent ni une expertise certifiée ni une
    décision de crédit. Authentification selon chaque endpoint, par X-API-Key ou Authorization: Bearer. Les
    endpoints authentifiés appliquent des quotas et peuvent retourner 429 avec Retry-After.
    Pour les clés API, les limites sont partagées par clé sur une même instance serveur : Free 10/minute
    et 200/24 heures, Pro 60/minute et 5000/24 heures, Enterprise 600/minute et 100000/24 heures.
    Les fenêtres démarrent au premier appel ; ces compteurs en mémoire ne constituent pas un quota
    global distribué. La clé sandbox partage le quota Free entre ses utilisateurs. Une indisponibilité
    de la vérification des clés peut retourner 503.
  contact:
    name: Support tevaxia.lu
    email: contact@tevaxia.lu
    url: https://www.tevaxia.lu
  license:
    name: Proprietary
servers:
  - url: https://tevaxia.lu
    description: Production
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
paths:
...
  /api/v1/ai/chat:
    post:
      tags:
        - AI
      summary: Conversation multi-tour avec l'assistant immobilier LU
      description: |
        Endpoint conversationnel spécialisé immobilier Luxembourg
        (fiscalité LIR, préparation de scénarios indicatifs, VEFA, copropriété, hôtellerie, KYC).
        Accepte jusqu'à 20 messages par requête.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messages
              properties:
                messages:
                  type: array
                  maxItems: 20
                  items:
                    type: object
                    required:
                      - role
                      - content
                    properties:
                      role:
                        type: string
                        enum:
                          - user
                          - assistant
                      content:
                        type: string
                        maxLength: 4000
      responses:
        '200':
          description: Réponse de l'assistant
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
                  model:
                    type: string
                  provider:
                    type: string
                  remaining:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
...